Remove `validation` module, refactor to `decorators`
